+++ tomcat/src/main/java/nextstep/jwp/vo/HttpCookie.java
@@ -1,14 +1,12 @@
 package nextstep.jwp.vo;
 
-import java.util.Collections;
 import java.util.Map;
 
-import static nextstep.jwp.vo.HttpHeader.JSESSION_ID;
+import static nextstep.jwp.vo.HeaderKey.JSESSION_ID;
 
 
 public class HttpCookie {
     private static final String NO_COOKIE = "";
-    private static final String COOKIE = "Cookie";
     private static final String COOKIE_DELIMITER = "; ";
 
     private final FormData cookies;
@@ -18,11 +16,11 @@ public class HttpCookie {
     }
 
     public static HttpCookie from(Map<String, String> headers) {
-        FormData formData = FormData.from(headers.getOrDefault(COOKIE, NO_COOKIE).split(COOKIE_DELIMITER));
+        FormData formData = FormData.from(headers.getOrDefault(HeaderKey.COOKIE.getName(), NO_COOKIE).split(COOKIE_DELIMITER));
         return new HttpCookie(formData);
     }
 
     public String getJsessionId() {
-        return this.cookies.get(JSESSION_ID.getValue());
+        return this.cookies.get(JSESSION_ID.getName());
     }
 }
