+++ tomcat/src/main/java/org/apache/coyote/http11/controller/DefaultController.java
@@ -8,7 +8,7 @@ import org.apache.coyote.http11.HtmlLoader;
 
 import java.io.IOException;
 
-import static nextstep.jwp.vo.HttpHeader.*;
+import static nextstep.jwp.vo.HeaderKey.*;
 
 public class DefaultController implements Controller {
 
@@ -22,9 +22,9 @@ public class DefaultController implements Controller {
         String responseBody = HtmlLoader.generateFile(PREFIX + fileName.concat());
 
         return Response.from(ResponseStatus.OK)
-                .addHeader(CONTENT_TYPE.getValue(),
-                        "text/" + fileName.getExtension() + CHARSET_UTF_8.getValue())
-                .addHeader(CONTENT_LENGTH.getValue(), String.valueOf(responseBody.getBytes().length))
+                .addHeader(CONTENT_TYPE.getName(),
+                        "text/" + fileName.getExtension() + CHARSET_UTF_8.getName())
+                .addHeader(CONTENT_LENGTH.getName(), String.valueOf(responseBody.getBytes().length))
                 .addBlankLine()
                 .addBody(responseBody);
     }
